perf: reduce to 1 slot

+++ src/MorphoV2.sol
@@ -75,22 +75,26 @@ contract MorphoV2 is IMorphoV2 {
         feeSetter = newFeeSetter;
     }
 
-    function setTradingFee(bytes32 id, uint128 tradingFee, uint128 interestCutLimit, bool activated) external {
+    function setTradingFee(bytes32 id, uint256 tradingFee, uint256 interestCutLimit, bool activated) external {
         require(msg.sender == feeSetter, "Only feeSetter");
         require(tradingFee <= WAD, "Trading fee too high");
         require(interestCutLimit <= WAD, "Interest cut limit too high");
-        tradingFeeParams[id] =
-            TradingFeeParams({tradingFee: tradingFee, interestCutLimit: interestCutLimit, activated: activated});
+        // Safe cast because values are below type(uint64).max.
+        tradingFeeParams[id] = TradingFeeParams({
+            tradingFee: uint64(tradingFee), interestCutLimit: uint64(interestCutLimit), activated: activated
+        });
     }
 
-    function setDefaultTradingFee(address loanToken, uint128 tradingFee, uint128 interestCutLimit, bool activated)
+    function setDefaultTradingFee(address loanToken, uint256 tradingFee, uint256 interestCutLimit, bool activated)
         external
     {
         require(msg.sender == feeSetter, "Only feeSetter");
         require(tradingFee <= WAD, "Trading fee too high");
         require(interestCutLimit <= WAD, "Interest cut limit too high");
-        defaultTradingFeeParams[loanToken] =
-            TradingFeeParams({tradingFee: tradingFee, interestCutLimit: interestCutLimit, activated: activated});
+        // Safe cast because values are below type(uint64).max.
+        defaultTradingFeeParams[loanToken] = TradingFeeParams({
+            tradingFee: uint64(tradingFee), interestCutLimit: uint64(interestCutLimit), activated: activated
+        });
     }
 
     function setTradingFeeRecipient(address recipient) external {
